temporarily enable MSM_TZ_LOG
The tz_log driver is responsible for allocating the memory for the
'qsee_log' buffer. If tz_log driver is disabled and some part of the
bootchain has set the qsee_log address before Linux booted, then that
previous address will continue being used. If that address overlaps
with memory used for some other purpose, logging to qsee_log may
scribble on memory and cause a crash.

The 'proper' solution for this will be to ensure that any users of
qsee_log in the bootchain (before Linux boots) unallocate their buffers
and clear the qsee_log pointer. Until support for that is implemented
in TZ and the bootloader, enable tz_log to avoid potential scribbling.

Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>